Create BinarySearch
This search algorithm needs the array to be sorted. As such, a basic sort method has been added to the tester class so that this algorithm can be used. This version uses the iterative version, using a while loop to sort the array. A recursive method can be implemented instead.