Update SearchAlgorithmTester
Added binary search. To use the multiple search algorithms, radio buttons have been implemented making use of the different search algorithm classes. Binary search only works on a sorted array, so a method has been added to sort the array (using insertion sort for simplicity). As this program is only supposed to focus on searching algorithms, the sorting algorithm is a rudimentary one that gets the job done. If larger arrays are implemented this method may need revisiting.